requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
python-telegram-bot==20.7
apscheduler==3.10.4
pydantic==1.10.15
python-dotenv==1.0.1
